requirements-file.txt
streamlit==1.30.0
openai==1.3.8
python-dotenv==1.0.0
nltk==3.8.1
